requirements.txt
Flask==3.1.0
flask-cors==5.0.1
pandas==2.2.3
groq==0.19.0
twikit==2.3.3
configparser==7.2.0
asyncio==3.4.3
uvicorn==0.34.0
gunicorn==23.0.0
requests==2.25.1
python-dotenv==1.0.1